Move migrated code into novoda package
