+++ java-library/src/main/java/uk/co/lucasweb/aws/v4/signer/Signer.java
@@ -12,6 +12,8 @@
  */
 package uk.co.lucasweb.aws.v4.signer;
 
+import com.novoda.aws.v4.signer.hash.Sha256;
+
 import javax.crypto.Mac;
 import javax.crypto.spec.SecretKeySpec;
 import java.nio.charset.Charset;
@@ -22,7 +24,6 @@ import java.util.List;
 import uk.co.lucasweb.aws.v4.signer.credentials.AwsCredentials;
 import uk.co.lucasweb.aws.v4.signer.credentials.AwsCredentialsProviderChain;
 import uk.co.lucasweb.aws.v4.signer.hash.Base16;
-import uk.co.lucasweb.aws.v4.signer.hash.Sha256;
 
 /**
  * @author Richard Lucas
