+++ java-library/src/test/java/uk/co/lucasweb/aws/v4/signer/AWSTestSuite.java
@@ -33,7 +33,7 @@ import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
 import uk.co.lucasweb.aws.v4.signer.credentials.AwsCredentials;
-import uk.co.lucasweb.aws.v4.signer.hash.Sha256;
+import com.novoda.aws.v4.signer.hash.Sha256;
 
 /**
  * @author Yoann Rodiere
