+++ kmp-library/src/commonMain/kotlin/com/novoda/aws/v4/signer/hash/StringExtension.kt
@@ -1,0 +1,3 @@
+package com.novoda.aws.v4.signer.hash
+
+expect fun String.toUtf8ByteArray(): ByteArray
