--- kmp-library/src/commonMain/kotlin/uk/co/lucasweb/aws/v4/signer/hash/StringExtension.kt
@@ -1,3 +1,0 @@
-package uk.co.lucasweb.aws.v4.signer.hash
-
-expect fun String.toUtf8ByteArray(): ByteArray
