--- kmp-library/src/iosMain/kotlin/uk/co/lucasweb/aws/v4/signer/hash/StringExtension.kt
@@ -1,3 +1,0 @@
-package uk.co.lucasweb.aws.v4.signer.hash
-
-actual fun String.toUtf8ByteArray(): ByteArray = toUtf8()
