+++ kmp-library/src/jvmMain/kotlin/com/novoda/aws/v4/signer/hash/StringExtension.kt
@@ -1,0 +1,3 @@
+package com.novoda.aws.v4.signer.hash
+
+actual fun String.toUtf8ByteArray(): ByteArray = toByteArray()
