--- kmp-library/src/jvmMain/kotlin/uk/co/lucasweb/aws/v4/signer/hash/StringExtension.kt
@@ -1,3 +1,0 @@
-package uk.co.lucasweb.aws.v4.signer.hash
-
-actual fun String.toUtf8ByteArray(): ByteArray = toByteArray()
